\name{CTT-package}
\alias{CTT-package}
\alias{CTT}
\docType{package}
\title{
Classical Test Theory Functions
}
\description{
This package can be used to perform a variety of tasks and analyses associated with classical test theory (CTT): 
score multiple-choice responses, perform reliability analyses, conduct item analyses, and transform scores onto 
different scales.}
\details{
\tabular{ll}{
Package: \tab CTT\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2014-02-26\cr
License: \tab GPL version 2 or newer\cr
}
The CTT package has the following functions: reliability, score, distractor.analysis, score.transform, spearman.brown, disattenuated.cor, subscales, polyserial.
}
\author{
John T. Willse <willse@uncg.edu>, Zhan Shu
}
\references{
     Crocker, L. & Algina, J. (1986). Introduction to Classical & Modern Test Theory, New York: Harcourt Brace Jovanovich College Publishers. \cr
     Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests. Psychometika, 16, 297-334. \cr
     Gulliksen, H. (1950). Theory of Mental Tests. New York: John Wiley & Sons, Inc. \cr
     Olsson, U., Drasgow, F. & Dorans, N. J. (1982). The Polyserial Correlation Coefficient. Psychometika, 47, 337-347. \cr

}
\keyword{ package }

