\name{CTTdata}
\alias{CTTdata}
\docType{data}
\title{ Example Multiple-Choice Data}
\description{
  This example data contains 20 unscored multiple-choice items that can be used with the CTT package.
}
\usage{data(CTTdata)}
\format{
  A data frame with 100 observations on the following 20 variables.
  \describe{
    \item{\code{i1}}{a character vector}
    \item{\code{i2}}{a character vector}
    \item{\code{i3}}{a character vector}
    \item{\code{i4}}{a character vector}
    \item{\code{i5}}{a character vector}
    \item{\code{i6}}{a character vector}
    \item{\code{i7}}{a character vector}
    \item{\code{i8}}{a character vector}
    \item{\code{i9}}{a character vector}
    \item{\code{i10}}{a character vector}
    \item{\code{i11}}{a character vector}
    \item{\code{i12}}{a character vector}
    \item{\code{i13}}{a character vector}
    \item{\code{i14}}{a character vector}
    \item{\code{i15}}{a character vector}
    \item{\code{i16}}{a character vector}
    \item{\code{i17}}{a character vector}
    \item{\code{i18}}{a character vector}
    \item{\code{i19}}{a character vector}
    \item{\code{i20}}{a character vector}
  }
}
\seealso{ CTTkey }
\examples{
data(CTTdata)
}
\keyword{datasets}
