\name{CTTkey}
\alias{CTTkey}
\docType{data}
\title{ Example Multiple-Choice Key}
\description{
  This example data contains a key for the 20 unscored multiple-choice items found in CTTdata and can be used with the CTT package.
}
\usage{data(CTTkey)}
\format{
  The format is:
 chr [1:20] "D" "C" "A" "D" "D" "A" "D" "B" "D" "A" ...
}
\seealso{ CTTdata }
\examples{
data(CTTkey)
}
\keyword{datasets}
